+++ src/main/java/app/AppBuilder.java
@@ -16,8 +16,7 @@ import interface_adapter.login.LoginViewModel;
 import interface_adapter.logout.LogoutController;
 import interface_adapter.logout.LogoutPresenter;
 import interface_adapter.logout.LogoutViewModel;
-import interface_adapter.reset.ResetController;
-import interface_adapter.reset.ResetPresenter;
+
 import interface_adapter.security.PasswordHasher;
 import interface_adapter.signup.SignupController;
 import interface_adapter.signup.SignupPresenter;
@@ -35,9 +34,6 @@ import use_case.login.LoginOutputBoundary;
 import use_case.logout.LogoutInputBoundary;
 import use_case.logout.LogoutInteractor;
 import use_case.logout.LogoutOutputBoundary;
-import use_case.reset.ResetInputBoundary;
-import use_case.reset.ResetInteractor;
-import use_case.reset.ResetOutBoundary;
 import use_case.service.UserService;
 import use_case.signup.SignupInputBoundary;
 import use_case.signup.SignupInteractor;
@@ -72,7 +68,7 @@ public class AppBuilder {
     private LogoutViewModel logoutViewModel;
     private LoginView loginView;
     private SignupView signupView;
-    private LogoutView logoutView;
+    private GameEndView logoutView;
     private GridView gridView;
     private InstructionsViewModel instructionsViewModel;
     private WordleInstructionsGUI wordleInstructionsGUI;
@@ -134,21 +130,13 @@ public class AppBuilder {
     }
 
 
-
-
-    public AppBuilder addLogoutView() {
+    public AppBuilder addGameEndView() {
         logoutViewModel = new LogoutViewModel();
-        logoutView = new LogoutView(logoutViewModel, userService);
+        logoutView = new GameEndView(logoutViewModel, userService);
         cardPanel.add(logoutView, logoutView.getViewName());
         return this;
     }
 
-    public AppBuilder addGameEndView() {
-        loginViewModel = new LoginViewModel();
-        loginView = new LoginView(loginViewModel);
-        cardPanel.add(loginView, loginView.getViewName());
-        return this;
-    }
 
     /**
      * Adds the Login Use Case to the application.
@@ -197,21 +185,13 @@ public class AppBuilder {
         final LogoutOutputBoundary logoutOutputBoundary = new LogoutPresenter(
                 viewManagerModel,
                 loginViewModel,
-                gridViewModel);
+                instructionsViewModel);
 
         final LogoutInputBoundary logoutInteractor =
                 new LogoutInteractor(userService, logoutOutputBoundary);
 
         final LogoutController logoutController = new LogoutController(logoutInteractor);
         logoutView.setLogoutController(logoutController);
-        return this;
-    }
-
-    public AppBuilder addResetUseCase() {
-        final ResetOutBoundary resetOutBoundary = new ResetPresenter(viewManagerModel, gridViewModel, instructionsViewModel);
-        final ResetInputBoundary resetInteractor = new ResetInteractor(resetOutBoundary);
-        final ResetController resetController = new ResetController(resetInteractor);
-        logoutView.setResetController(resetController);
         return this;
     }
