--- src/main/java/interface_adapter/reset/ResetState.java
@@ -1,4 +1,0 @@
-package interface_adapter.reset;
-
-public class ResetState {
-}
