--- src/main/java/interface_adapter/reset/ResetViewModel.java
@@ -1,4 +1,0 @@
-package interface_adapter.reset;
-
-public class ResetViewModel {
-}
