--- src/main/java/use_case/reset/ResetInputBoundary.java
@@ -1,7 +1,0 @@
-package use_case.reset;
-
-public interface ResetInputBoundary {
-    void execute(ResetInputData resetInputData);
-
-    void switchToInstructionView();
-}
