--- src/main/java/use_case/reset/ResetInputData.java
@@ -1,13 +1,0 @@
-package use_case.reset;
-
-public class ResetInputData {
-    private final String username;
-
-    public ResetInputData(String username) {
-        this.username = username;
-    }
-
-    String getUsername() {
-        return username;
-    }
-}
