--- src/main/java/use_case/reset/ResetOutBoundary.java
@@ -1,8 +1,0 @@
-package use_case.reset;
-
-public interface ResetOutBoundary {
-
-    void prepareSuccessView(ResetOutputData resetOutputData);
-
-    void switchToInstructionView();
-}
